MuralPicker: Abort previous requests when selecting a workspace or room
Abort previous requests when selecting a workspace or room. This fixes
an issue where the picker unexpectedly shows data for the previously
selected workspace after the user picks a different workspace.

The issue occurred when the requests from the user's previous selection
complete after the requests from the user's most recent selection. This
was likely to happen when the previously selected workspace contained a
large number of rooms and murals and the most recently selected
workspace contained a small number of rooms and murals.